Modified How Cards are Passed Around
Added a "discard pile" to Deck objects to and now the "burn cards" are
pushed to that array instead of calling array.shift() by itself. This
isn't much use for this app, but I am trying to make the code work for a
range of different games.

Deck can now deal out one card at a time. Table can now receive one card
at a time. And, although completely unnecessary, the deck is dealt in a
way that is consistent with handling a deck in real life. This means
that cards are burnt (sent to the discard pile) in between steps, like
dealing to the table and then to a player.