Fixed Error in when checking for Straight Flush
If a hand had a card that was higher than the maximum value of a
sequence of cards, checkStraightFlush would return null, even if a
Straight Flush was contained within the cards given.

This was cause by one of the checks being (counter > 4) when it should
be (counterMax > 4).

After making the change, checkStraightFlush now returns the proper data.